# Remove usage of 'nonMetricNameCharRE' regexp