Another optimization made by removing usage of regexp. To keep PRs small, I'm removing only one of the three regexps here and will open more for the others.

Benchmark results for `BuildCompliantMetricName`:
```
arthursens$ benchstat main.txt no-regexp.txt 
goos: darwin
goarch: arm64
pkg: github.com/prometheus/otlptranslator
cpu: Apple M2 Pro
                                                                                      │   main.txt   │            no-regexp.txt            │
                                                                                      │    sec/op    │    sec/op     vs base               │
BuildCompliantMetricName/withSuffixes=true/Basic_metric_with_no_special_characters-2     869.5n ± 2%   183.4n ±  3%  -78.91% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Counter_metric-2                              958.1n ± 1%   233.0n ±  5%  -75.69% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Gauge_ratio_metric-2                          857.3n ± 1%   224.5n ±  1%  -73.81% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_with_per-unit_suffix_notation-2        997.9n ± 3%   322.9n ±  2%  -67.64% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_with_special_characters-2             1473.0n ± 1%   238.7n ± 10%  -83.79% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_starting_with_digit-2                  656.6n ± 0%   159.5n ±  1%  -75.71% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_with_multiple_underscores-2           1928.5n ± 0%   264.9n ±  1%  -86.26% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_with_complex_unit-2                   1470.5n ± 2%   455.0n ±  1%  -69.06% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=false/Basic_metric_with_no_special_characters-2    746.9n ± 1%   741.6n ±  0%   -0.71% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=false/Counter_metric-2                             800.7n ± 1%   794.8n ±  1%   -0.74% (p=0.041 n=6)
BuildCompliantMetricName/withSuffixes=false/Gauge_ratio_metric-2                         689.7n ± 1%   706.0n ±  4%        ~ (p=0.084 n=6)
BuildCompliantMetricName/withSuffixes=false/Metric_with_per-unit_suffix_notation-2       469.4n ± 1%   478.5n ±  3%        ~ (p=0.065 n=6)
BuildCompliantMetricName/withSuffixes=false/Metric_with_special_characters-2             1.439µ ± 4%   1.445µ ±  1%        ~ (p=0.582 n=6)
BuildCompliantMetricName/withSuffixes=false/Metric_starting_with_digit-2                 544.8n ± 1%   541.9n ±  2%        ~ (p=0.310 n=6)
BuildCompliantMetricName/withSuffixes=false/Metric_with_multiple_underscores-2           1.974µ ± 1%   1.991µ ±  3%        ~ (p=0.260 n=6)
BuildCompliantMetricName/withSuffixes=false/Metric_with_complex_unit-2                   696.1n ± 1%   693.1n ±  1%        ~ (p=0.121 n=6)
geomean                                                                                  945.4n        452.2n        -52.17%

                                                                                      │  main.txt  │            no-regexp.txt            │
                                                                                      │    B/op    │    B/op     vs base                 │
BuildCompliantMetricName/withSuffixes=true/Basic_metric_with_no_special_characters-2    244.0 ± 0%   160.0 ± 0%  -34.43% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Counter_metric-2                             376.0 ± 0%   288.0 ± 0%  -23.40% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Gauge_ratio_metric-2                         288.0 ± 0%   208.0 ± 0%  -27.78% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_with_per-unit_suffix_notation-2       360.0 ± 0%   256.0 ± 0%  -28.89% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_with_special_characters-2             373.0 ± 0%   240.0 ± 0%  -35.66% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_starting_with_digit-2                 188.0 ± 0%   120.0 ± 0%  -36.17% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_with_multiple_underscores-2           413.0 ± 0%   240.0 ± 0%  -41.89% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_with_complex_unit-2                   521.0 ± 0%   368.0 ± 0%  -29.37% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=false/Basic_metric_with_no_special_characters-2   100.0 ± 0%   100.0 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Counter_metric-2                            104.0 ± 0%   104.0 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Gauge_ratio_metric-2                        96.00 ± 0%   96.00 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Metric_with_per-unit_suffix_notation-2      72.00 ± 0%   72.00 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Metric_with_special_characters-2            244.0 ± 0%   244.0 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Metric_starting_with_digit-2                76.00 ± 0%   76.00 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Metric_with_multiple_underscores-2          220.0 ± 0%   220.0 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Metric_with_complex_unit-2                  96.00 ± 0%   96.00 ± 0%        ~ (p=1.000 n=6) ¹
geomean                                                                                 194.0        159.5       -17.80%
¹ all samples are equal

                                                                                      │  main.txt   │            no-regexp.txt            │
                                                                                      │  allocs/op  │ allocs/op   vs base                 │
BuildCompliantMetricName/withSuffixes=true/Basic_metric_with_no_special_characters-2    21.000 ± 0%   6.000 ± 0%  -71.43% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Counter_metric-2                             24.000 ± 0%   8.000 ± 0%  -66.67% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Gauge_ratio_metric-2                         21.000 ± 0%   7.000 ± 0%  -66.67% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_with_per-unit_suffix_notation-2        23.00 ± 0%   10.00 ± 0%  -56.52% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_with_special_characters-2             33.000 ± 0%   6.000 ± 0%  -81.82% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_starting_with_digit-2                 17.000 ± 0%   6.000 ± 0%  -64.71% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_with_multiple_underscores-2           43.000 ± 0%   6.000 ± 0%  -86.05% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=true/Metric_with_complex_unit-2                    31.00 ± 0%   12.00 ± 0%  -61.29% (p=0.002 n=6)
BuildCompliantMetricName/withSuffixes=false/Basic_metric_with_no_special_characters-2    15.00 ± 0%   15.00 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Counter_metric-2                             16.00 ± 0%   16.00 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Gauge_ratio_metric-2                         14.00 ± 0%   14.00 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Metric_with_per-unit_suffix_notation-2       10.00 ± 0%   10.00 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Metric_with_special_characters-2             28.00 ± 0%   28.00 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Metric_starting_with_digit-2                 11.00 ± 0%   11.00 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Metric_with_multiple_underscores-2           37.00 ± 0%   37.00 ± 0%        ~ (p=1.000 n=6) ¹
BuildCompliantMetricName/withSuffixes=false/Metric_with_complex_unit-2                   14.00 ± 0%   14.00 ± 0%        ~ (p=1.000 n=6) ¹
geomean                                                                                  20.50        11.01       -46.30%
¹ all samples are equal
```

Benchmark results for `BuildMetricName`:

```
benchstat main.txt no-regexp.txt 
goos: darwin
goarch: arm64
pkg: github.com/prometheus/otlptranslator
cpu: Apple M2 Pro
                                                                             │  main.txt   │           no-regexp.txt            │
                                                                             │   sec/op    │    sec/op     vs base              │
BuildMetricName/withSuffixes=true/Basic_metric_with_no_special_characters-2    42.41n ± 1%   42.56n ±  2%       ~ (p=0.368 n=6)
BuildMetricName/withSuffixes=true/Counter_metric-2                             65.92n ± 1%   64.59n ±  1%  -2.00% (p=0.004 n=6)
BuildMetricName/withSuffixes=true/Gauge_ratio_metric-2                         81.95n ± 3%   81.83n ±  1%       ~ (p=1.000 n=6)
BuildMetricName/withSuffixes=true/Metric_with_per-unit_suffix_notation-2       117.3n ± 6%   116.9n ±  3%       ~ (p=0.781 n=6)
BuildMetricName/withSuffixes=true/Metric_with_special_characters-2             45.46n ± 4%   44.17n ±  1%  -2.83% (p=0.015 n=6)
BuildMetricName/withSuffixes=true/Metric_starting_with_digit-2                 43.55n ± 3%   41.87n ±  2%  -3.87% (p=0.009 n=6)
BuildMetricName/withSuffixes=true/Metric_with_multiple_underscores-2           47.66n ± 7%   45.04n ±  0%  -5.51% (p=0.002 n=6)
BuildMetricName/withSuffixes=true/Metric_with_complex_unit-2                   162.0n ± 2%   155.9n ±  1%  -3.74% (p=0.002 n=6)
BuildMetricName/withSuffixes=false/Basic_metric_with_no_special_characters-2   24.79n ± 1%   24.23n ±  0%  -2.26% (p=0.002 n=6)
BuildMetricName/withSuffixes=false/Counter_metric-2                            24.71n ± 1%   24.11n ±  1%  -2.41% (p=0.002 n=6)
BuildMetricName/withSuffixes=false/Gauge_ratio_metric-2                        26.56n ± 7%   24.15n ±  2%  -9.05% (p=0.002 n=6)
BuildMetricName/withSuffixes=false/Metric_with_per-unit_suffix_notation-2      25.94n ± 4%   23.70n ±  2%  -8.62% (p=0.002 n=6)
BuildMetricName/withSuffixes=false/Metric_with_special_characters-2            26.93n ± 1%   25.95n ±  1%  -3.62% (p=0.002 n=6)
BuildMetricName/withSuffixes=false/Metric_starting_with_digit-2                24.22n ± 2%   23.67n ± 30%       ~ (p=0.121 n=6)
BuildMetricName/withSuffixes=false/Metric_with_multiple_underscores-2          27.01n ± 3%   27.38n ± 13%       ~ (p=0.240 n=6)
BuildMetricName/withSuffixes=false/Metric_with_complex_unit-2                  24.30n ± 3%   24.25n ±  3%       ~ (p=1.000 n=6)
geomean                                                                        41.39n        40.20n        -2.87%

                                                                             │  main.txt  │           no-regexp.txt            │
                                                                             │    B/op    │    B/op     vs base                │
BuildMetricName/withSuffixes=true/Basic_metric_with_no_special_characters-2    48.00 ± 0%   48.00 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Counter_metric-2                             96.00 ± 0%   96.00 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Gauge_ratio_metric-2                         112.0 ± 0%   112.0 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Metric_with_per-unit_suffix_notation-2       120.0 ± 0%   120.0 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Metric_with_special_characters-2             64.00 ± 0%   64.00 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Metric_starting_with_digit-2                 40.00 ± 0%   40.00 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Metric_with_multiple_underscores-2           80.00 ± 0%   80.00 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Metric_with_complex_unit-2                   176.0 ± 0%   176.0 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Basic_metric_with_no_special_characters-2   32.00 ± 0%   32.00 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Counter_metric-2                            32.00 ± 0%   32.00 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Gauge_ratio_metric-2                        32.00 ± 0%   32.00 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Metric_with_per-unit_suffix_notation-2      24.00 ± 0%   24.00 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Metric_with_special_characters-2            48.00 ± 0%   48.00 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Metric_starting_with_digit-2                24.00 ± 0%   24.00 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Metric_with_multiple_underscores-2          64.00 ± 0%   64.00 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Metric_with_complex_unit-2                  32.00 ± 0%   32.00 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                                                        53.22        53.22       +0.00%
¹ all samples are equal

                                                                             │  main.txt  │           no-regexp.txt            │
                                                                             │ allocs/op  │ allocs/op   vs base                │
BuildMetricName/withSuffixes=true/Basic_metric_with_no_special_characters-2    2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Counter_metric-2                             3.000 ± 0%   3.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Gauge_ratio_metric-2                         3.000 ± 0%   3.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Metric_with_per-unit_suffix_notation-2       4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Metric_with_special_characters-2             2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Metric_starting_with_digit-2                 2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Metric_with_multiple_underscores-2           2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=true/Metric_with_complex_unit-2                   5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Basic_metric_with_no_special_characters-2   1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Counter_metric-2                            1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Gauge_ratio_metric-2                        1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Metric_with_per-unit_suffix_notation-2      1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Metric_with_special_characters-2            1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Metric_starting_with_digit-2                1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Metric_with_multiple_underscores-2          1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=6) ¹
BuildMetricName/withSuffixes=false/Metric_with_complex_unit-2                  1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                                                        1.645        1.645       +0.00%
¹ all samples are equal
```